requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
hashids==1.3.1
sqlparse==0.5.5
tzdata==2025.3
python-dotenv==1.0.1
stripe==12.5.1

